renamed reader to parser as gives a better meaning for the job it does
